requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
huggingface-hub==0.30.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.4.2
kaleido==0.2.1
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.7.2
mpmath==1.3.0
narwhals==1.33.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.24.3
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==10.2.0
plotly==6.0.1
pynndescent==0.5.13
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==4.0.2
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
typing_extensions==4.13.1
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.3.0
wordcloud==1.9.4
zipp==3.21.0
google-genai
python-dotenv
pygraphviz  # Optional, for advanced layouts
